segementing out click commands
